requirements.txt
Pyrogram==2.0.96
pytelegrambotapi==4.9.0
python-dotenv==1.0.0
